# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Punjab Blood Connect API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /hospital/profile:
    get:
      operationId: getHospitalProfile
      tags: [hospital]
      summary: Get own hospital profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Hospital profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HospitalProfile"
        "401":
          description: Unauthorized

    patch:
      operationId: updateHospitalProfile
      tags: [hospital]
      summary: Update own hospital profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HospitalProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HospitalProfile"
        "401":
          description: Unauthorized
  /hospital/inventory:
    get:
      operationId: getHospitalInventory
      tags: [hospital]
      summary: Get own blood inventory
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Blood inventory
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BloodInventoryItem"
        "401":
          description: Unauthorized
  /hospital/inventory/{bloodGroup}:
    patch:
      operationId: updateBloodStock
      tags: [hospital]
      summary: Update blood stock for a specific blood group
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: bloodGroup
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BloodStockUpdate"
      responses:
        "200":
          description: Updated inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodInventoryItem"
        "401":
          description: Unauthorized
  /hospital/inventory/history:
    get:
      operationId: getInventoryHistory
      tags: [hospital]
      summary: Get blood stock update history
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Stock history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockHistoryItem"
        "401":
          description: Unauthorized
  /hospital/change-password:
    post:
      operationId: changeHospitalPassword
      tags: [hospital]
      summary: Change hospital password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthMessage"
        "400":
          description: Validation error
        "401":
          description: Unauthorized

  # --- ADMIN ---
...
  /admin/categories:
    get:
      operationId: adminListCategories
      tags: [admin]
      summary: List all categories
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

    post:
      operationId: adminCreateCategory
      tags: [admin]
      summary: Create a category
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /admin/categories/{id}:
    patch:
      operationId: adminUpdateCategory
      tags: [admin]
      summary: Update a category
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

    delete:
      operationId: adminDeleteCategory
      tags: [admin]
      summary: Delete a category
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BloodInventoryItem:
      type: object
      required: [id, hospitalId, bloodGroup, units, status, lastUpdated]
      properties:
        id: { type: integer }
        hospitalId: { type: integer }
        bloodGroup: { type: string }
        units: { type: integer }
        status: { type: string }
        lastUpdated: { type: string }
    StockHistoryItem:
      type: object
      required: [id, hospitalId, bloodGroup, unitsBefore, unitsAfter, action, createdAt]
      properties:
        id: { type: integer }
        hospitalId: { type: integer }
        bloodGroup: { type: string }
        unitsBefore: { type: integer }
        unitsAfter: { type: integer }
        action: { type: string }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
...
    Category:
      type: object
      required: [id, name, isActive, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: ["string", "null"] }
        color: { type: ["string", "null"] }
        isActive: { type: boolean }
        createdAt: { type: string }
    CategoryInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: string }
        color: { type: string }
    CategoryUpdate:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        color: { type: string }
        isActive: { type: boolean }

    # --- PUBLIC ---
...
    AuthMessage:
      type: object
      required: [message]
      properties:
        message: { type: string }
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string }
        newPassword: { type: string }

    # --- HOSPITAL SELF-REGISTRATION ---
...
    HospitalProfile:
      type: object
      required: [id, name, type, ownership, city, district, address, contactNumber, status]
      properties:
        id: { type: integer }
        name: { type: string }
        registrationNumber: { type: string }
        type: { type: string }
        ownership: { type: string }
        city: { type: string }
        district: { type: string }
        state: { type: string }
        address: { type: string }
        pinCode: { type: ["string", "null"] }
        contactPerson: { type: string }
        contactNumber: { type: string }
        email: { type: string }
        website: { type: ["string", "null"] }
        workingHours: { type: ["string", "null"] }
        googleMapsLat: { type: ["number", "null"] }
        googleMapsLng: { type: ["number", "null"] }
        logoUrl: { type: ["string", "null"] }
        approvalStatus: { type: string }
        status: { type: string }
        createdAt: { type: string }
    HospitalProfileUpdate:
      type: object
      properties:
        contactPerson: { type: string }
        contactNumber: { type: string }
        address: { type: string }
        pinCode: { type: string }
        website: { type: string }
        workingHours: { type: string }
        googleMapsLat: { type: number }
        googleMapsLng: { type: number }
    BloodStockUpdate:
      type: object
      required: [units]
      properties:
        units: { type: integer }
        notes: { type: string }

    # --- ADMIN HOSPITAL ---
...