# lib/api-zod/src/generated/types/categoryUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Punjab Blood Connect API
 * OpenAPI spec version: 0.2.0
 */

export interface CategoryUpdate {
  name?: string;
  description?: string;
  color?: string;
  isActive?: boolean;
}